pkg/macaroon: intersect allowed buckets instead of union
Caveats in macaroons must only allow them to become more restrictive
because anyone is allowed to add a caveat while it is impossible to
remove them. Since we were doing a union of all of the allowed paths
to find the allowed buckets, someone could get information leaks
about which buckets existed by adding new path restrictions and
listing buckets.

This changes the code to perform an intersection of the set of
buckets for each caveat, and to return an error if any of the
caveats disallow the action (though that should have already been
handled by an earlier call to Check).

Change-Id: Ia69c56dde9c74ba9897a75999999fc7c26f0ca0e