rpc/{rpcstatus,rpctimeout}: wrap rpc errors with stacks and add timeout package
the rpctimeout package allows grpc/drpc servers to perform stream
operations with a timeout. since the only way to cancel a stream
operation on the server side is to exit from the handler, we have
to launch a goroutine to issue the operation, return early if the
timeout is hit, and then leak the goroutine and pray it eventually
cleans up after we return from the handler. i hate it, but it's
the documented way to do things.

Change-Id: I55fac6d60f7b6683dd42cf786f1d31c8983b9dcf