rpc: Add a function to resolve a storage node address to an IP address.
Added a function named LookupNodeAddress() that takes a storage node address and resolves it to the first IP address resolved to play safe with DNS load balancing. If an IP address is provided that IP address is returned back which makes calling this function tolerant of any storage node address.
Change-Id: I47f19923c2bc1f4a3bc398183b02d73d4f09a03e