pkcrypto: Fix PublicKeyFromPrivate func signature
The PublicKeyFromPrivate function returned a private key and an error
indistinctly due to cypto.PrivateKey type is an empty interface.

Returning an error as a crypto.PrivateKey is confusing and the only
explanation for it is that it was in that way due to a bug.

This commit change the signature of the function and update all the
callers of this repository that call it.

Change-Id: I7680e628cd46eba335fd1dc5f2e820e27b0e6afe