strictcsv: simple and strict csv marshaling/unmarshaling
The upcoming storage node compensation reporting work relies on
generating and ingesting CSV files. There are existing packages, like
github.com/gocarina/gocsv but unfortunately they are very lenient by
default and the knobs to make them more strict are package level and
not per-parser, making them unideal.

This is a simple package that marshals CSV to/from structs or slices of
structs. It requires:
1) all fields be annotated with a `csv` tag
2) all fields map to CSV headers
3) all CSV headers map to fields

Change-Id: Ife1951184d18ef45e21d1299480c86f2d03ecd7b