encryption: fix key derivation with bucket base
bases added to the store are expected to already have the
bucket name folded in to the key derivation, but our check
to see if we needed to fold in the bucket name had false
positives. it was looking to see if we were using the default
key (and hence needed to fold in the bucket name) by checking
if we consumed any from the path passed in to the store. this
works in all cases except when the path is the empty path. in
that case, the bucket name would be folded in twice, causing
incorrect keys to be derived. fix this false positive by adding
a Default field to the base.

Change-Id: I95d46601546f68cf2f354870bb143d91151098ea