encryption: add EncryptPrefixWithStoreCipher
consider the case where a user puts to the keys

    foo/bar
    foo/bar/
    foo/bar/baz

intuitively, if you were to list with the prefix "foo/bar/" one
might expect to get results like {"", "baz"}. similarly, one
might expect a restriction to that prefix to be able to access
exactly those (not foo/bar).

as it currently is, becasue we tend to add and remove slashes
in some places, sharing would give you access to all three
files. i have no idea what listing would do, but i expect that
if it gets this one right, it gets another one wrong.

the problem is that encrypting a path is a different operation
than encrypting a prefix. consider the above paths' encryption
where enc() is encrypting the empty string:

    enc(foo)/enc(bar)
    enc(foo)/enc(bar)/enc()
    enc(foo)/enc(bar)/enc(baz)

this is why we sometimes remove the trailing slash when sharing.
if we did not, a share for "foo/bar/" would only share the key
for "foo/bar/" and not "foo/bar/baz", since the permissions happen
on the encrypted form.

so to fix this, we add encrypt and decrypt prefix functions that
take into account the trailing slash. this means that if you
called EncryptPrefix on the paths, you would have

    enc(foo)/enc(bar)
    enc(foo)/enc(bar)/
    enc(foo)/enc(bar)/enc(baz)

which keeps listing and sharing working the way one would hope.

we do not need a corresponding DecryptPrefix method because the
decryption code already handles leaving trailing slashes. by that,
i mean that decrypting "enc(foo)/enc(bar)/" gives "foo/bar/".

many more changes will probably be coming to things like the
encryption store, libuplink implementation, and metainfo services
to work, but this is the first step.

Change-Id: Ic977b2311f912fed2009259b11e5b997d5d69066