rpc/rpccache: introduce cache type for connections
it's a bounded LRU allowing multiple entries per key with
bounds based both on total size and on number of entries
per key. it also has expiration and support for checking if
entries are stale, as determined by the value, before returning
them.

rather than tie it to specific drpc types or rpc types, it
just uses interface{} for keys and values. this allows it to
potentially be reused in other situations that have the same
criteria.

Change-Id: I7df698b1a1fb47e9831970fb592f3686be554a31