telemetry/tracing: log enabled state instead of disabled
Telemetry and tracing are opt-in services. Hence users would be more
interested to see a log message when they are enabled, instead of only
having a message when they are disabled.

Also, all logs with INFO and higher severity are user-facing and they
should be formatted well and consistent with other logs, i.e. they
should start with a capital letter and have sentence style.

Change-Id: I4550509c59b5ec5e44eefe36d85d8183105f6a6d