rpc: better connection pooling
this changes it so that the dialer has a connection pool on it
so that dials to the same node with the same dialer have an
opportunity to be reused. this is mostly expected to benefit
satellites.

it's opt in because you have to call NewDefaultPooledDialer
in order to get the pool behavior.

Change-Id: I92fb6cad7075cb6e50551ae000f963bc62c707e4