common/socket: qos background dialer
the purpose of this change is to provide a library to ease the use
of the new uplink.Config DialContext field. example usage:

  cfg := uplink.Config{
    DialContext: socket.BackgroundDialer().DialContext,
  }

This will automatically set QoS flags and congestion controller
settings to be background traffic, if possible.

Storj breaks all other devices on my home network without this,
and not all of our users are able to change their router settings.

Change-Id: Ia6e30b2a3d1790947d8fa9ee2d65e0877597b9d6