private/process: support hardcoding metric application name/suffix
using argv[0] or letting the application class be configured
was a mistake. our storage node rothko data has application
classes all over the place, not to mention the storagenode vs
storagenode.exe problem.

this allows certain programs to opt in to hardcoding the
application name, storagenode for example

i'd like to fix satellites too but that will change all the
satellite metric names

Change-Id: I2afcbe966984dec32c9d1cb87ba31387b219f218