process: add func Report to trigger telemetry client reporting
We would like to add telemetry to the zombie segment reaper so we can
graph the amount of garbage we observe. However, the segment reaper is
not a long-lived process. The telemetry client reports data on an interval,
and the segment reaper process may end before the data is sent.

Change-Id: Idb09bb15e7e3941c26056e2471eef9b1a86fa73a