rpc: refactor rpc.Transport to also handle tls
This is part 1 of our QUIC effort.

Since QUIC automatically does tls handshake when establishing a connection, we need to make our
rpc.Transport layer to be able to handle tls, so we don't have to rewrite a lot of Dialer level logic.

This PR moves tls related logic into rpc.Transport layer.

Change-Id: I7ac935334a4c877e9e92794a3770134d374a804f