process: fill in Use only if empty
Many processes fill it up already with things like "uplink"
and we now have code that sets the default log formatter
based on it. Without this change, it would be filled to the
full path of the binary, like "/home/jeff/go/bin/uplink"
which is not what we want.

Change-Id: I6e7894982ca25287657a0c103a593e62eb3370a3