cfgstruct: error out if a default value set for an internal flag
Currently, there's no indication that the value set with `default:"true"` will
not be applied if `internal: "true"` is specified.
It would be nice to know such behavior during `bind` call.
For example:
type Config struct {
  EnableA `help: "enable A service" internal:"true" default:"true"`
}
In this case, the code intends to set `EnableA` by default to true, but the
current behavior would set `EnableA` to false. If `default` value is not honored,
we should error out instead.

Change-Id: I6d6fc575a87c5f2a56a21184c7338ed56a4de081