pb: Remove an unintentional deprecated option
The deprecated option was added unintentionally in a previous commit
when we removed the order Settlement service function and despite was
detected by a reviewer it may be added back during a conflicting rebase
and then was overlooked by the author and reviewers.

This commit drops unintentional the deprecated option.

Change-Id: Id86a793ea5c7cdcbd8d490b2a836e0c12c437ddb