pb: Update SegmentListResponse to include Range
This change updates the comments on ObjectDownloadRequest to state that
the returned SegmentListItems only include those within the requested
range. It also adds a Range request to the SegmentListRequest so that,
when making a SegmentListRequest you can specify a plain range. This
allows a client who needs `more` items to include the necessary info.

It also adds a plain_offset to the SegmentListItem, so that the client
knows which portion of the plain data is covered by this segment.

Change-Id: Ice9d7a764cc32a87ea1ef86b4d6d8c0e179eddf3