common/rpc/rpcpool: handle drpc idle connection closes
This solution fixes a wide variety of "uplink: metainfo error: EOF"
bugs. It fixes most of them and should be safe to deploy to
production.

We in general have an idempotence problem where as soon as bytes
start going over the wire, we can't really know if we raced with
a connection close on the server side, if we get an error on Write.

However, we could reduce some windows of race by having drpc tell
us if the connection was closed on an Invoke call, instead of
checking first.

This change also disables the server-side inactivity timeout again
for now until more clients have this client-side change.

Change-Id: I46d3fa7d19ad5f2adf8bc6f9a04f4e291fee6d3c