rpc/rpcpool: fix panic with nil pool and add basic test
a refactoring to avoid duplicating the logic to extract the
tls connection state caused the code to panic when used with
a nil pool, which is how "no connection pooling" is supposed
to work. this fixes that and adds a minimal test to ensure
that it at least performs some basic duties with respect to
caching and avoiding pooling.

Change-Id: I76ec6f3062a40ce26d4f70a96c1f2fede831b16c