version: allow printing of build tags
Currently our version package does not print out -rc when calling
the String() method on the SemVer. This creates confusion when
running an build like v1.42.0-rc in testing, as it currently reports
incorrectly as v1.42.0 only.

Furthermore it also currently breaks the autoupdater, since our
versioncontrol support the build tag handout, and the updater gets
confused when downloading the right binary which then reports
incorrectly.

Change-Id: Ie5d98371b54765d23f5dc97dc65c92dcccd6ce93