rpc: change the way we do quic rollouts
this fixes three problems with the quic rollout.

1. we were always dialing and doing a tls handshake to the peer
   even if quic was disabled. this had to be happening since
   we could compute the peer identity

2. it was implemented by having the quic dialer itself return
   an error if the rollout decides to not use quic. but that
   caused a number of side problems. errors were polluted which
   caused problems with audits. directly using the quic
   connector like the satellite does when pinging nodes for
   quic status was erroring. probably more stuff.

3. the quick percent override was a global thing which is
   problematic for tests in general. global state bad, yet
   again.

instead, this changes the quic rollout to be computed before
dialing, and if the rollout decides to not use quic, we use
the context to instruct the hybrid connector to force a
tcp dial. additionally, we use the context to influence the
rollout percent (thanks egon).

Change-Id: I39619156af692ed9a4815d2ca6b61eb7519699de