storj: allow JSON encoding of OrderLimits
JSON encoding of OrderLimits is used by the satellite fetch-pieces
command already, but there is a problem: the output produced contains
"{}" for each uplink_public_key field, because we have provided the json
machinery no way to encode or decode the PiecePublicKey type.

With this change, OrderLimits can be completely serialized to JSON and
deserialized while still remaining validatable.

A test is added here also so that this quality should continue to hold
in the face of future changes.

Change-Id: I73d982cf76c9a6ffb2971bc7d96abd1244e1f76e