debug: Add descriptions to list links index page
The debug server serves an index page with links to several endpoints
that provide different type of "debug/monitoring" information.

This commit add a description to each one so the user can know what to
expect to find on them.

The previous rendered page looked like:

  /version/
  /debug/pprof/
  /debug/pprof/symbol
  /debug/run/trace/db
  /mon/
  /metrics
  /health
  /logging

With these changes, the rendered page looks like:

  /version/
     JSON representation of the current version information for the binary
  /debug/pprof/
     HTML page listing the available Go pprof profiles
  /debug/pprof/symbol
     Looks up the program counters listed in the request, responding with a table mapping program counters to function names
  /debug/run/trace/db
     Collects traces until request is canceled
  /mon/
     HTML page listing the available Monkit metrics
  /metrics
     Monkit metrics in Prometheus format
  /health
     Debug server health check

Change-Id: I75501d8a4356cf990c598ef894b8a9e8e4967f00