rpc: lazy dialing
we do a lot of unnecessary dials, especially when we don't have
connection pooling enabled, so that we can get access to the
peer id before we potentially do any communication with them.
this is problematic for performance and other things (like
sending initial 0-rtt data).

this changes the returned connections from the dialer to be
lazy dialing instead of immediately dialing and putting the
connection back into the pool.

it is probably very disruptive and will need some testing.
it may even be prudent to change the PeerIdentity API to
have a different name so that we are forced to audit all
uses.

Change-Id: I3f380e56d1c765dc50f054fbcfd9aa841ad4625c