quic rollout: stay at 0 for now
we built this rollout functionality but then
discovered that due to a bug
(https://review.dev.storj.io/c/storj/common/+/7096),
we did rollout QUIC already in a large number of cases.

in chronograf:

   SELECT cumulative_sum(sum("delta")) AS "conns"
   FROM "v3_stats_new"."autogen"."incoming_connection"
   WHERE time > now() - 1h
   GROUP BY time(:interval:), "kind" FILL(null)

it appears quic is already handling 1/3rd of all satellite
connections. because of the nature of the bug, that means
storage node dials are also getting 1/3 of connections with
quic attempts.

to start, we want to start the rollout with satellite
quic only.

we are not yet confident about the performance of quic in
parallel settings.

Change-Id: If6f960bcd0106d210eec6090c99ed039c428f166