common/socket: use env var for lowPrioCongControl
CAIA Delay-Gradient (CDG) is similar to ledbat but has the advantage of
being in-kernel. For Scavenger traffic (Less-than-Best-Effort) set the
module parameters use_shadow=0 and use_ineff=0

for root users, specifying cdg by default might result in an unnexpected
change in behavior, so we're controlling this with an env var that defaults
to nothing.

closes https://github.com/storj/storj/issues/4606

Change-Id: Iaadd07c49935892c397681d98cc604830b27861f