process: add ExecWithCustomConfigAndLogger function
As an attempt to improve the logging of the storagenode and storagenode updater processes
within the new storagenode docker image (which houses both the storagenode and updater binaries),
we need a new ExecX function in the process package that takes a custom logger function as argument.

This change adds ExecWithCustomConfigAndLogger function.

Usage:

    func main() {
        loggerFunc := func(logger *zap.Logger) (*zap.Logger) {
            return logger.With(zap.String("Process", "storagenode-updater"))
        }

        process.ExecWithCustomConfigAndLogger(rootCmd, true, process.LoadConfig, loggerFunc)
    }

Updates https://github.com/storj/storj/issues/4648

Change-Id: Ib8dc36897f85295971d33d68e36450d469f9faed