sync2: implement ReadCache
ReadCache allows to cache a read result. It has threshold for triggering
a background refresh and for requiring waiting for a result.

We need logic to improve satellite/overlay/downloadselection and
satellite/overlay/uploadselection, such that refreshing doesn't
block until actually necessary. This implementation allows background
refreshes and allows them to be reused.

Change-Id: Ic1c7319f42555ea2ae6c9783caf83d2c45ef53e4