process: make the env prefix configurable
Since 7e63afbef (and following commits) we have one container for storagenode which also starts storagenode-updater.

Which is awesome, as the storagenode will be automatically updated by the updater, inside the container.

Unfortunatelly it breaks the possibility to monitor the storagenode instance with setting `STORJ_DEBUG_ADDR=0.0.0.0:11111`.

Because STORJ_DEBUG_ADDR is used both by storagenode-updater AND storagenode and they couldn't use the same port. :-(

This makes it hard to monitor storagenode in containerized environments.

As a workaround I would make the configuration prefix (`STORJ_...`) configurable, and modify the entrypoint to use different prefix for storagenode-updated.

This would make it possible to use `STORJ_` environment variables only for the `storagenode` process (but still makes it possible to configure `storagenode-updater`).

Change-Id: I879fa96c2e06c106292e0c229a21466edfe6ac63