sync2: improvements for limiter
* ensure we never start a new goroutine when the context has been cancelled.
* ensure every goroutine has finished after call to wait and that no
  new goroutine can be started after wait has been called.

Currently it would be easy to introduce a mistake when using a Limiter
inside of a server.

One of the critical requirements was that `.Wait` must be called only if
there is no potential for `.Go` call. This is the same approach that
sync.WaitGroup and errgroup.Group use. However, it's also an easy
mistake to make.

Instead remove some flexibility from the Limiter, such that it cannot be
reused for waiting. This ensures that when Wait / Close returns it's
guaranteed that there won't be more goroutines started.

Also use "Close" instead of "Wait", because it makes it obvious that it
cannot be reused at a later date.

Change-Id: Ie67149ebcc42911149c69dd95e90ecd2c5c7be52