experiment: propagate feature flag between layers
There are some interesting questions, which can be answered only in production.

Example: the obligatory file sync (used by the piecestore writer of the storagenode) might be too slow on a significant part of storage nodes.

We need to do some measurement without a final implementation change.

For this reason we introduce a protocol level feature flag to turn on and off such kind of experiments.

IMPORTANT: it's designed to be used only for temporary experiments. We wouldn't like to use it for real feature flags (like QUIC enabled/disabled). If sg. is a real (even optional) production feature, should add it to the method signatures to make them more explicit (instead of hiding them inside the context values).

This is only for temporary experiments to avoid refactoring if they will be reverted anyway...

The implementation is very simple to the propagation of trace information. It's saved to the drpcmetadata for outgoing drpc messages, and saved back to the context when the messages arrive (see export/import.go)

Change-Id: I4034a825b83f56addd02b97032cf87bdc926e586