pb: metainfo; rename EncryptedPath to EncryptedObjectKey
In the past we were using inaccurate naming because we don't really have
files and paths but we have objects and keys. We started to change it and most of the system is using correct naming but we have some leftovers and metainfo protocol is one example. This change will also force changes to storj/uplink and storj/storj repositories.

https://github.com/storj/uplink/issues/59

Change-Id: I049318e096bc0e132e8e81601078dc09a9a559bf