pb: add desired_nodes to download requests
During the download the satellite tries to find the optimal number of storagenodes to download from.

In case of default 29/80/110 RS scheme, it's 39.

There are specific use cases where client needs to request more nodes. This option can be used to download faster from nodes, with also paying for more egress.

Change-Id: Ic6b02cddaeb9b4f84682ac8ccb157a13101bb73a