version,process: Revert "add application name and instance to version_info monkit metric"
This reverts commit 7c84ef61baa540ddd70a728effabe522f2d1d6f9.

Reason for revert: application and instance are already in the monkit metrics as part of the monkit delivery packet. this is doubling the data sent over the wire, and increasing cardinality unnecessarily (again, these values are already in the monkit data stream). i will follow up to make sure we solve the problem we have in a better way.

References https://github.com/storj/storj/issues/5255

Change-Id: I170616fd44b8f8d21ec2260bd89660ee548af945