versions: add age metric (time since last commit)
This patch extends `version_info` metric with the time, since the commit, used for the build.
(assuming go>=1.18 is used, where vcs.time build information is added to the binary)

This is an important metric, shows the agility of the CI/CD pipeline.

For example:
 * in case of a 2 weeks release cycle,
 * where release is immediatelly deployed,
 * it's sg between 0 and 2 weeks (1 week on average).

 * With 2 weeks of release cycle
 * + 1 week testing in QA,
 * it's sg. between 1 and 3 weeks (average 1.5)

If this value is less, the implemented features are deployed more quickly after the merge.

Change-Id: Ia8e6b3fb65903a11b5409e2130bdc58dfabb8a15