peertls/tlsopts: disable distributed tracing for verifyIdentity
verifyIdentity uses default, un-parented context, therefore it doesn't make sense to include in distributed traces, it just increases the noise.

To avoid circurlar dependency, this requires to move the excluded.go to a separated package (tracing), but for backward-compatility, a deprecated version is also kept.

Majority of this commit is just package rename. The real change is in `peertls/tlsopts/tls.go` (2 lines).

https://github.com/storj/storj/issues/5722

Change-Id: I6b48f788998adb414f0bd86f6a434fbb6eb1214d