process: add ExecWithCustomOptions function
As part of working on https://github.com/storj/storj/issues/5349
to run separate services with low I/O Priority for garbage collection
and used-space calculation, we want to be able to disable the profiler
debug server, tracing and other systems already enabled in the main
process.

This change adds an ExecWithCustomOptions function to the process
package to give us the option to disable the aforementioned systems.

Change-Id: I9baf1339449ab6622a9df61c2c305398f0269daf