pb/piecestore: add max chunk size parameter to downloads
right now, nodes always send back exactly the amount
the uplinks requested for the chunck, which seems fine
but has some negative effects.

    1. the chunks are large which cause larger allocations
       adversely impacting peak memory usage.
    2. because it's in a protobuf wrapper, we can't process
       the first part of the chunk without fully downloading
       the chunk, hurting our time to first byte.
    3. the throttle is not as granular as it could be because
       of the large sizes, causing stalls and fairness issues.
    4. the uplink logic has a difficult time pipelining new
       allocation requests because it has no insight into when
       the current allocation is nearing completion, which means
       that the round trip latency is paid for each chunk,
       reducing throughput.

it's hard to know what a good maximum chunk size should be
without doing some emperical tests, but that's hard to
do when the parameter is configured node side. instead of
that, the nodes will enforce a maximum maximum, and then
the uplinks can request a maximum chunk size allowing
them to tune for their own needs.

Change-Id: I38e29d5cc72f049c24e047941581be0b70be255f