process: force zap logger to use RFC3339 time encoder
Currently, with the time encoder set to ISO8601, zap logger
uses different ISO8601 time formats across different platforms.
On windows, it uses the datetime with time offset format
`2023‐06‐29T13:12:39+00:00` which is not recognized by
time.UnmarshalJSON because it parses time using strictly
RFC3339.

This change changes the time encoder of the zaplogger to use
RFC3339 encoder.

Resolves https://github.com/storj/storj/issues/6006

Change-Id: Ife33826d0dcef260cea9732e2fed136355662b48