process/exec_conf: fix error handling of the configuration loading
Today we have different behavior, depends on the configuration source.

Let's say we have an invalid configuration value (value, not key!).

 1. If it's set by a flag --> process startup is failed.
 2. If it's set by an environment variable --> problem is logged on INFO level and ignored.
 3. If it's set by a configuration value --> problem is logged on INFO level and ignored.

Probably, because we use configuration files (and environment variables) which are shared between multiple microservices.

But there is an unforgivabble type of error, where the key value is bad. Like a wrong NodeID or a wrong placement configuration.

In this case, it's more safe to fail as early as possible.

But to avoid breaking storagenodes, only a configuration option is introduced in this patch, to use the stricter validation for satellites (not for storagenodes).

https://github.com/storj/storj/issues/6211

Change-Id: Id0649476498d77682aae4e9b25a0e8ae7ed67568