metrics: introduce restricted /metrics server
This patch provides a helper to start HTTP servers which can expose all the monkit metrics in Prometheus format.

Existing `debug` package also includes this functionality, but it also has additional endpoints, which are mode dangerous to expose.

The metrics server is:

 * exposes only the read-only `/metrics` endpoint
 * available only via TLS + client cert

This patch can help the monitoring of storagenode fleets in various environments, as storagenodes can be monitored with per-host Prometheus instance (or without additional VPN configuration to access metrics).

Change-Id: I90538c11af4cf941d61a4ba2c44f28cf3f0fe058