{grant,encryption}: limit restricted enc access from a cloned parent
When restricting an access grant, the underlying encryption access
should be limited to the restricted API key from a cloned copy of the
parent access.

Before this change, the restricted encryption access was created from
scratch, which missed to limit it to the allowed paths of any previous
caveats in the macaroon chain.

Change-Id: I08973e5d8f54e341869765aa0ffbc2c6212d8655