rpc/rpcpool: lifetime option for connections
We have IdleExpiration time for the connection pool, but that's time is reset when we re-use the connection.

Therefore it's possible to use a very-very old connection, if it's not idle, but used frequently.

In some scenarios, we may need to re-open the connection to avoid strange network errors.

Opening one connection per 5-10 minutes it's not a big deal from performance point of view.

Change-Id: I9015e0cc58ce920dd8fe91115740b240adbb2fa1