common/storj: remove DownloadNodes from RedundancyScheme
this method is used to calculate what the long tail should
be for downloads from the long tail prescribed for uploads.

unfortunately, these should not always be the same and it
is desired we can configure upload and download long tails
separately.

initially, i considered adding another field to the
RedundancyScheme to support specifying this, but then i
discovered that there are only two places DownloadNodes
is used -

  * in the satellite, to determine how many order limits
    to generate and
  * in the uplink, which limits the number of order limits
    it uses to this number.

we can remove the uplink use entirely because it can
just use the satellite's choice. it doesn't need to limit
the value to itself. see
https://review.dev.storj.io/c/storj/uplink/+/12258

so we're going to move this logic (and subsequent
configuration) to the satellite. see
https://review.dev.storj.io/c/storj/storj/+/12259

Change-Id: I41bbffe5a666676a480cec34a81a02871e763b5d