sync2: improve WithTimeout
the old code would launch a goroutine for every
WithTimeout call. the new code instead uses
the time.AfterFunc function to avoid a goroutine
because the runtime keeps timers in a separate
heap allowing it to be more efficient. we still
need to make a channel and coordinate with the
timer because AfterFunc launches the function
in a new goroutine when it does fire.

another side benefit of more rarely launching
a goroutine is that runtime/trace traces have
many less goroutines and events to keep track
of. this function is used for every read/write
call for downloads so it's a big win.

benchmark results:

              | old         | new         | delta
    sec/op    | 650.1n ± 3% | 139.4n ± 2% | -78.55%
    alloc/op  | 440         | 200         | -54.55%
    allocs/op | 6           | 3           | -50.00%

Change-Id: Idff469a91446b9ea760607a58a67138803f4c8b5