pb: add method for retrieving bucket's Object Lock configuration
The metainfo RPC service has been given a new method
GetBucketObjectLockConfiguration that returns a bucket's Object Lock
configuration. Currently, the configuration only indicates whether a
bucket has Object Lock enabled, but in the future, it will be extended
to include the default retention mode and retention period duration for
new objects once we allow those settings to be configured.

Change-Id: I562ff39282a7c6eb658589d35364f1d81af04896