storj,pb: adjust retention mode constants
This change removes the retention mode constants relating to legal hold
because users of this package should not treat an object's legal hold
status as a property of its retention mode. Instead, legal hold status
will be represented as a separate property in structures holding object
information.

Additionally, the value of the retention mode constant for governance
mode has been changed from 3 to 2.

Change-Id: I19ec8a503a6a1cd1538136f44a85765d62357fb7