pb: make object legal hold status nullable
The object protobuf definition contains a field indicating an object's
legal hold status. This field was typed as a regular boolean because
it was thought that we would only need to indicate whether legal hold
was enabled or disabled. However, in order to be consistent with Amazon
S3, we must be able to represent a third state indicating that the
legal hold status cannot be discerned due to insufficient permissions.
For this reason, the type of the legal hold field has been changed.

Change-Id: I96550386fa722635492899d169ad4ddfdae36a26