accesslogs: move the package from storj.io/edge
This change moves the accesslogs package to this repository as it might
be used by others teams/services in the future other than edge.

It doesn't move s3(_test).go because that's exclusive to edge services
and StorjUploader to prevent cyclic dependency between this package and
libuplink (but it's not a big chunk of code anyway).

Updates storj/storj-private#1079

Change-Id: I1c9ce157b5f00434c6ef3c52034710e3adf71f15