pb: add if_none_match field to support conditional writes
This adds a new repeated string `if_none_match` field to the
following:

* `CommitObjectRequest`
* `FinishCopyObjectRequest`

We use a repeated string to match closely with the HTTP header.
However, for the first version of the implementation we will only
be checking for a single string of '*' to indicate an object
existence check should be performed, and any other value to not
perform the check.

In the future we may add support to specify etags for
`If-None-Match` and `PUT` requests, which could be useful to
validate the latest object version, not just that the object exists.

Note: There is no `If-None-Match` field when creating a new
multi-part upload in S3, so setting the field on `BeginObject`
would only apply to `PutObject` requests. To be more consistent
with S3 we only provide `if_none_match` for commit requests, which
is used by both upload methods.

Reference:

https://docs.aws.amazon.com/AmazonS3/latest/userguide/conditional-writes.html

Closes storj/edge#561

Change-Id: I877ec057f42708ebfb82a5c05f5640c5d395dfe0