rpc/rpccache: close filtered stale entries
when a conn is expired, we also check all of the
other entries in the cache for that key and filter
them if they are stale. a stale entry may not be
closed once we added support for considering
connections stale if they are too old and not
closed by the remote or by a cancel or something.

so, call close and hopefully we don't leak
connections anymore.

Change-Id: Ieac9d6075976d2b1e2b9ceb263f26e1ec8f5dbc7