rpc/rpccache: use linked list
a linked list gives O(1) deletions and insertions
into the global order which outscales slices in
the hundreds of entries based on benchmarks run
at the time of the original commit:

https://review.dev.storj.tools/c/storj/common/+/1697/comment/1973a4ec_5ee8a588/

at that time it was expected that we would have
hundreds of entries in the cache, but it turns out
we have servers now with tens of thousands of
entries in the cache, and almost all of the time
is spent in an expensive pointer-aware memmove
under a mutex. a linked list is appropriate in
this case.

this also fixes a couple of potential bugs:

1. when a value expires, we called filterCacheKey
which filtered all of the entries for the key if
they were stale but this depended on stale being
immediately set after close. that may not be the
case. instead we can just filter the entry that
we know is closed directly.

2. if an error in the NewStream call happens in
the rpcpool, we didn't replace the entry in the
cache which would leak the pool value. it's
unlikely that an error in NewStream would happen
without the connection being closed, but that
doesn't mean it's impossible.

Change-Id: I2a20cf8ed6ee8fe624a0d965ea90e65b3738f656