all: Homogenize zap fields
We have a disperse manners of writing the name of Zap logger field
names.

This commit homogenize all of them to use snake_case with only lowercase
ASCII letters, numbers and underscores.

Ignore the log field names of the package process/gcloudlogging which
must follow the format of the Google Cloud Logging because they are in
purpose.

This is a companion commit for the new linter that we'll detect Zap
field names which don't follow this convention
https: //review.dev.storj.tools/c/storj/ci/+/19816

Change-Id: Ib3ca78b60295563a6c522f2b8695263c728a67ba